added registers

+++ src/cpu.h
@@ -3,4 +3,4 @@ extern uint16_t program_counter;
 extern uint16_t addr_select;
 
 bool cycle();
-void init_cpu(uint16_t memsize);
+void init_cpu(int memsize);
